sandbox-runtime/pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sandbox-runtime"
version = "0.1.0"
authors = [
  { name="Chen Xiao", email="xavier.chen@aishu.cn" },
]
description = "Sandbox environment for code execution"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]
dependencies = [
    "aiohttp==3.12.0",
    "fastapi==0.115.12",
    "paramiko==3.5.1",
    "psutil==7.1.3",
    "pydantic==2.11.5",
    "pydantic-settings==2.12.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.1",
    "requests==2.32.3",
    "uvicorn==0.34.3",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "python-multipart>=0.0.6",
    "python-multipart>=0.0.20",
]
sdk = [
    "requests>=2.31.0",
    "aiohttp>=3.12.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
sandbox-run = "sandbox_runtime.cli.main:entry_point"


[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
